pyproject.toml
[project]
name = "fusor"
description = "Computable object representation and validation for gene fusions"
authors = [
    {name = "Alex Wagner", email = "alex.wagner@nationwidechildrens.org"},
    {name = "Kori Kuzma", email = "kori.kuzma@nationwidechildrens.org"},
    {name = "James Stevenson", email = "james.stevenson@nationwidechildrens.org"},
    {name = "Katie Stahl", email = "kathryn.stahl@nationwidechildrens.org"},
    {name = "Jeremy Arbesfeld", email = "jeremy.arbesfeld@nationwidechildrens.org"}
]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
license = {file = "LICENSE"}
dependencies = [
    "pydantic == 2.*",
    "ga4gh.vrsatile.pydantic ~=0.2.0",
    "ga4gh.vrs ~=0.8.1",
    "biocommons.seqrepo",
    "gene-normalizer ~=0.1.40-dev1",
    "cool-seq-tool ~=0.3.0-dev1",
]
dynamic=["version"]

[project.optional-dependencies]
tests = ["pytest", "pytest-cov", "pytest-asyncio"]
dev = ["pre-commit", "pytest", "pytest-cov", "pytest-asyncio", "ruff>=0.1.14"]

[project.urls]
Homepage = "https://github.com/cancervariants/fusor"
Documentation = "https://github.com/cancervariants/fusor"
Changelog = "https://github.com/cancervariants/fusor/releases"
Source = "https://github.com/cancervariants/fusor"
"Bug Tracker" = "https://github.com/cancervariants/fusor/issues"


[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "fusor.version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"fusor.examples" = ["*.json"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true

[tool.ruff]
src = ["src"]
select = [
    "F",  # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "E", "W",  # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "I",  # https://docs.astral.sh/ruff/rules/#isort-i
    "N",  # https://docs.astral.sh/ruff/rules/#pep8-naming-n
    "D",  # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "UP",  # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "ANN",  # https://docs.astral.sh/ruff/rules/#flake8-annotations-ann
    "ASYNC",  # https://docs.astral.sh/ruff/rules/#flake8-async-async
    "S",  # https://docs.astral.sh/ruff/rules/#flake8-bandit-s
    "B",  # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "A",  # https://docs.astral.sh/ruff/rules/#flake8-builtins-a
    "C4",  # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "DTZ",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
    "T10",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
    "EM",  # https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
    "G",  # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
    "PIE",  # https://docs.astral.sh/ruff/rules/#flake8-pie-pie
    "T20",  # https://docs.astral.sh/ruff/rules/#flake8-print-t20
    "PT",  # https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
    "Q",  # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    "RSE",  # https://docs.astral.sh/ruff/rules/#flake8-raise-rse
    "RET",  # https://docs.astral.sh/ruff/rules/#flake8-return-ret
    "SIM",  # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
    "PTH",  # https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
    "PGH",  # https://docs.astral.sh/ruff/rules/#pygrep-hooks-pgh
    "RUF",  # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
]
fixable = [
    "I",
    "F401",
    "D",
    "UP",
    "ANN",
    "B",
    "C4",
    "G",
    "PIE",
    "PT",
    "RSE",
    "SIM",
    "RUF"
]
# ANN101 - missing-type-self
# ANN003 - missing-type-kwargs
# D203 - one-blank-line-before-class
# D205 - blank-line-after-summary
# D206 - indent-with-spaces*
# D213 - multi-line-summary-second-line
# D300 - triple-single-quotes*
# D400 - ends-in-period
# D415 - ends-in-punctuation
# E111 - indentation-with-invalid-multiple*
# E114 - indentation-with-invalid-multiple-comment*
# E117 - over-indented*
# E501 - line-too-long*
# W191 - tab-indentation*
# *ignored for compatibility with formatter
ignore = [
    "ANN101", "ANN003",
    "D203", "D205", "D206", "D213", "D300", "D400", "D415",
    "E111", "E114", "E117", "E501",
    "W191"
]

[tool.ruff.per-file-ignores]
# ANN001 - missing-type-function-argument
# ANN2 - missing-return-type
# ANN102 - missing-type-cls
# N805 - invalid-first-argument-name-for-method
# F401 - unused-import
# S101 - assert
"tests/*" = ["ANN001", "ANN2", "ANN102", "S101"]
"*__init__.py" = ["F401"]
"src/fusor/models.py" = ["ANN201", "N805", "ANN001", "ANN2"]
